Se añade ajuste de fecha, se añaden textos informaivos en consola y se ajustan los recusros que tendrá que mover el usuario es decir el archivo donde moverá la fecha
