+++ POM/data/posBMR/constantesPosBMR_Decrepado.ts
@@ -1,0 +1,635 @@
+export const REPORTE_POSBMR = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    PLATAFORMA2: 'Captura Abono',
+    PLATAFORMA3: '06 Chedraui',
+    PLATAFORMA4: '12 Liverpool',
+    PLATAFORMA5: 'WalMart',
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+
+    INTERNACIONAL: '0 Domestico',
+
+    EMISOR: '00 Todas',
+
+    EMISOR2: '0 Master y Visa',
+    
+    GRUPO_CADENA: '',
+
+    //Todas - T | Compras - C | Devoluciones - D
+    TIPO_TRANSACCION: 'T',
+
+    //Reporte Original - O | Desgloce Compras / Devoluciones - D
+    TIPO_REPORTE: 'D',
+
+    CODIGO_RECHAZO: 'Ninguno',
+
+    FECHA_INICIO:'24012025' ,
+
+    FECHA_FIN: '24012025',
+
+    SALIDA_EXCEL : true
+    
+}
+
+export const REPORTE_1TOTAL_TRANC_ACEPTADAS = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_2DETALLE_TRANC_ACEPTADAS = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_3TOTAL_TRANC_RECHAZADAS = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_4DETALLE_TRANC_RECHAZADAS = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_5RECHAZAS_NO_PROCESADAS = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+ 
+    //Escribir el numero
+    VENTANA: '1',
+    
+    
+}
+
+export const REPORTE_6CONSOLIDADOS_RECHAZADOS_X_FECHA = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+
+    //Escribir el numero
+    VENTANA: '1',
+    
+}
+
+export const REPORTE_7TOTALES_EMISOR = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    PLATAFORMA2: 'Captura Abono',
+    PLATAFORMA3: '06 Chedraui',
+    PLATAFORMA4: '12 Liverpool',
+    PLATAFORMA5: 'WalMart',
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+   
+}
+
+export const REPORTE_8DETALLE_EMISOR = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    PLATAFORMA2: 'Captura Abono',
+    PLATAFORMA3: '06 Chedraui',
+    PLATAFORMA4: '12 Liverpool',
+    PLATAFORMA5: 'WalMart',
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+   
+}
+
+export const REPORTE_9CONSOLIDADO_EMISOR = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    PLATAFORMA2: 'Captura Abono',
+    PLATAFORMA3: '06 Chedraui',
+    PLATAFORMA4: '12 Liverpool',
+    PLATAFORMA5: 'WalMart',
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+   
+}
+
+export const REPORTE_ACONSOLIDADO_PLATAFORMA = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+   
+}
+
+export const REPORTE_BRETENCION_RIESGO = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    INTERNACIONAL: '0 Domestico',
+
+}
+
+export const REPORTE_CLIBERACION_RIESGO = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    EMISOR: '00 Todas',
+    
+}
+
+export const REPORTE_DCONSOLIDADO_RIESGO_INT = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+
+    EMISOR2: '0 Master y Visa', 
+}
+
+export const REPORTE_EDET_RIESGO_INT = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    EMISOR2: '0 Master y Visa', 
+}
+
+export const REPORTE_FTXN_SIS_CAP = {
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+}
+
+export const REPORTE_GRECH_LOTES_CAPTURA = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA2: 'Captura Abono',
+  
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+
+}
+
+export const REPORTE_HTOT_TXN_PROMOCIONES = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA2: 'Captura Abono',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+     
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_I_DET_TXN_PROMOCIONES = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA2: 'Captura Abono',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+     
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_JDET_TXN_RECH_PROMOCIONES = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA2: 'Captura Abono',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+     
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_KTOT_TXN_ACEP_GRUPO_CADENA = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //Escribir tal cual viene en el select 
+    TRANSACCION: null,
+
+    GRUPO_CADENA: '',
+
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+
+    OPCIONMODAL: "Totales 1",
+}
+
+export const REPORTE_LTOT_TXN_RECH_GRUPO_CADENA = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA: '7Eleven',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    GRUPO_CADENA: '',
+
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+
+    OPCIONMODAL: "Totales 1",
+      
+}
+
+export const REPORTE_MCONSOLIDADO_PROMOCIONES = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Escribir tal cual viene en el select
+    PLATAFORMA3: '06 Chedraui',
+    
+}
+
+export const REPORTE_NSUMARIO_PROMOCIONES_BANAMEX = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA4: '12 Liverpool',
+  
+    //Todas - T | Compras - C | Devoluciones - D
+    TIPO_TRANSACCION: 'T',
+
+    //Reporte Original - O | Desgloce Compras / Devoluciones - D
+    TIPO_REPORTE: 'D',
+    
+}
+
+export const REPORTE_PREPORTE_WALMART = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA5: 'WalMart',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //Escribir el numero
+    VENTANA: '1',
+    
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+}
+
+export const REPORTE_QTRANSACCIONES = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+
+    //Pesos = P | Dolares= D, siempre en mayuscula
+    MONEDA: 'P',
+    
+    //Escribir tal cual viene en el select
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+}
+
+
+export const REPORTE_RREPORTE_PAGOS_CASH = {
+    
+    //DIA/MES/AÑO 
+    FECHA:'24012025',
+  
+    //Escribir tal cual viene en el select
+    PLATAFORMA6: 'Oxxo',
+    
+    //Escribir num afiliacion
+    AFILIACION: '',
+
+    //Escribir num tarjeta
+    TARJETA: '',
+    
+    //true=se activa el check | false = se desactiva el check
+    SUBTOTALES: false,
+    
+}
+
+export const REPORTE_SREPORTE_RECHAZOS= {
+    
+    //DIA/MES/AÑO 
+    FECHA_INICIO:'24012025' ,
+    CODIGO_RECHAZO: 'Ninguno',
+}
+
+
+export const REPORTE_VAPUNTOS_BBVA = {
+    
+    FECHA_INICIO:'24012025' ,
+
+    FECHA_FIN: '24012025',
+
+    SALIDA_EXCEL : false
+    
+}
