+++ Recursos/POSBMR.ts
@@ -1,0 +1,1 @@
+export const FECHA = "24012025";
